[A2.2.0] +handle_memeber_event handle_memeber_event is a private helper function that handles the member callbacks. this is more just to make the parts of the code that are public facing more easy to read.
+set_game_pad_button_callback
sets the call back for the game_pad_button events  note that there is only one callback that it calls for both the key up and down events.

*changed dispactch_SDL_event to do_event and simplified it
originally dispactch_SDL_event pumped the event loop itself when I rewrote it to not do that I left some code that checked to see if the event was handled (because I was orginally only returning from the function when it was either break or quit) so that I could flag if the event was handled or not. Now that I am expecting the user to manually pump the event cue I can just return early if I have a callback to the function.
I renamed it to do event because the orignal name is too long and also don't want to limit the function to only doning sdl related stuff.