[A14.0.0] added a simple C++ini file parser
since win16's ini parsing function is depercated (because you are supposed to use the registry instead) and not cross platform I reimplemented it in C++ using c++20's ranges
Also added a number concept that aliases isNumber.
and a type_traits sub namespace to concepts and a is_number type trait for the number conecept.