[A22.12.0] added simple random number generator class
added SKC::GE::RNG which is a simple class that is a layer of abstraction over the standard lib's random number generators.
SKC::GE::RNG is a template class that takes a concepts::number (integral or floating point) type that is the output type and a std::uniform_random_bit_generator type that is the random source.